feat(fleet-console): give Session Analyst the Quick Launch model chip and three-rung effort track
Move EffortTrack onto the plugin SDK so the analyst composer can share Quick Launch's chip and track grammar. Clamp analyst effort to low, medium, and high.
