+++ .changelog.d/analyst-ql-track-proposal.md
@@ -1,0 +1,8 @@
+---
+branch: analyst-ql-track-proposal
+---
+
+### fleet-console
+#### Changed
+- Session Analyst now picks its model and effort with the same chip and three-rung track as Quick Launch, limited to low, medium, and high.
+  ko: Session Analyst의 모델·강도를 Quick Launch와 같은 칩과 3단 트랙으로 고르게 하고, 강도는 low·medium·high만 엽니다.
