+++ runtime/fleet-console/core/client/src/main.tsx
@@ -6,6 +6,7 @@ import * as sdkOperationsBrowser from "@fleet-console/sdk/operations/browser";
 import * as sdkNotificationsBrowser from "@fleet-console/sdk/notifications/browser";
 import * as sdkReactBrowser from "@fleet-console/sdk/react/browser";
 import * as sdkComponentsFailureNotice from "@fleet-console/sdk/components/failure-notice";
+import * as sdkComponentsEffortTrack from "@fleet-console/sdk/components/effort-track";
 import "@fontsource-variable/fraunces";
 import "@fontsource-variable/fraunces/standard-italic.css";
 import "@fontsource-variable/manrope";
@@ -40,6 +41,7 @@ interface FleetConsoleRuntime {
   readonly "@fleet-console/sdk/notifications/browser": typeof sdkNotificationsBrowser;
   readonly "@fleet-console/sdk/react/browser": typeof sdkReactBrowser;
   readonly "@fleet-console/sdk/components/failure-notice": typeof sdkComponentsFailureNotice;
+  readonly "@fleet-console/sdk/components/effort-track": typeof sdkComponentsEffortTrack;
 }
 
 declare global {
@@ -57,6 +59,7 @@ globalThis.__fleetConsoleRuntime__ = {
   // shim은 이 네임스페이스에서 실제 모듈을 읽는다 — 정의만 등록하고 여기를 비워 두면
   // 외부 플러그인이 로드되는 순간 "runtime shim unavailable"로 죽는다.
   "@fleet-console/sdk/components/failure-notice": sdkComponentsFailureNotice,
+  "@fleet-console/sdk/components/effort-track": sdkComponentsEffortTrack,
 };
 
 // 서버 주입이 첫 페인트의 권위값이며 theme-hint는 미주입 서빙 경로의 폴백이다.
