+++ runtime/fleet-plugins/terminal/client/i18n/index.ts
@@ -69,7 +69,10 @@ export const terminalEn = {
   "terminal.analyst.initialSettings": "Initial analysis settings",
   "terminal.analyst.cli": "Analysis CLI",
   "terminal.analyst.model": "Analysis model",
+  "terminal.analyst.modelMenu": "Analysis models",
   "terminal.analyst.effort": "Analysis effort",
+  "terminal.analyst.effortAuto": "AUTO",
+  "terminal.analyst.effortAutoValue": "Automatic",
   "terminal.analyst.na": "n/a",
   "terminal.analyst.composerPlaceholder": "Ask about the session… (/ for commands)",
   "terminal.analyst.stop": "Stop",
@@ -585,7 +588,10 @@ export const terminalKo: Record<keyof typeof terminalEn, string> = {
   "terminal.analyst.initialSettings": "초기 분석 설정",
   "terminal.analyst.cli": "분석 CLI",
   "terminal.analyst.model": "분석 모델",
+  "terminal.analyst.modelMenu": "분석 모델",
   "terminal.analyst.effort": "분석 Effort",
+  "terminal.analyst.effortAuto": "자동",
+  "terminal.analyst.effortAutoValue": "자동",
   "terminal.analyst.na": "n/a",
   "terminal.analyst.composerPlaceholder": "세션에 대해 질문하기… (/ 명령)",
   "terminal.analyst.stop": "중지",
