+++ scripts/generate-fleet-console-shim-keys.mjs
@@ -34,6 +34,7 @@ const SPECIFIER_ENTRIES = [
   ["@fleet-console/sdk/notifications/browser", path.join(sdkRoot, "notifications", "browser.ts")],
   ["@fleet-console/sdk/react/browser", path.join(sdkRoot, "react", "browser.tsx")],
   ["@fleet-console/sdk/components/failure-notice", path.join(sdkRoot, "components", "failure-notice.tsx")],
+  ["@fleet-console/sdk/components/effort-track", path.join(sdkRoot, "components", "effort-track.tsx")],
 ];
 
 const tmpDir = mkdtempSync(path.join(os.tmpdir(), "fleet-console-shim-keys-"));
